requirements.txt

PyYAML>=6.0
requests>=2.31.0
graphql-core>=3.2.3
python-dotenv>=1.0.0
coverage==7.8.0

